requirements-test.txt
affine==2.2.1
amqp==2.3.2
appdirs==1.4.3
astroid==2.0.4
attrs==19.1.0
beautifulsoup4==4.6.3
billiard==3.5.0.4
black==18.9b0
boltons==18.0.1
bs4==0.0.1
cachetools==3.0.0
celery==4.2.1
certifi==2018.10.15
cf-units==2.0.2
cftime==1.0.2.1
chardet==3.0.4
checksumdir==1.1.4
ciso8601==2.1.1
Click==7.0
click-plugins==1.0.4
cligj==0.5.0
cloudpickle==0.6.1
colorama==0.4.0
compliance-checker==4.1.1
coverage==4.5.2
coveralls==1.5.1
cssselect==1.0.3
Cython==0.29
dask==0.20.2
dataclasses==0.6
DAWG==0.7.8
docopt==0.6.2
docutils==0.14
editdistance==0.5.2
fake-useragent==0.1.11
Fiona==1.8.2
Flask==1.0.2
Flask-Caching==1.4.0
GDAL==2.4.0
GeoAlchemy2==0.5.0
geographiclib==1.49
idna==2.7
isodate==0.6.0
isort==4.3.15
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==3.0.1
kombu==4.2.1
lazy-object-proxy==1.3.1
lxml==4.2.5
MarkupSafe==1.1.0
mccabe==0.6.1
mock==2.0.0
munch==2.3.2
netCDF4==1.4.2
nose==1.3.7
numpy==1.13.3
OWSLib==0.17.0
pandas==0.23.4
parse==1.9.0
pbr==3.1.1
pendulum==2.0.4
pep8==1.7.1
pluggy==0.9.0
psycopg2==2.7.6.1
py==1.5.2
pydash==4.7.4
pyee==5.0.0
pyflakes==2.1.1
pygeoif==0.7
pylint==1.7.0
pyorbital==1.5.0
pyparsing==2.3.0
pyPEG2==2.15.2
pyppeteer==0.0.25
pyproj==1.9.5.1
pyquery==1.4.0
pytest==4.3.1
pytest-cov==2.6.1
pytest-isort==0.3.1
python-dateutil==2.7.5
python-rapidjson==0.6.3
pytz==2018.7
pytzdata==2018.7
PyYAML==3.13
rasterio==1.0.10
redis==3.0.1
regex==2017.7.28
requests==2.21.0
requests-html==0.9.0
scipy==1.1.0
Shapely==1.6.4.post2
simplekml==1.3.1
singledispatch==3.4.0.3
six==1.11.0
snuggs==1.4.2
SQLAlchemy==1.2.14
structlog==18.2.0
tabulate==0.8.2
toolz==0.9.0
tqdm==4.28.1
typed-ast==1.1.0
urllib3==1.24.1
vine==1.1.4
w3lib==1.19.0
websockets==7.0
Werkzeug==0.14.1
wrapt==1.10.11
xarray==0.11.0
Werkzeug==0.14.1
wrapt==1.10.11
xarray==0.11.0
